ui: create unique component only for private key deleting
